requirements.txt
pandas
fastapi
python-docx
